Adding 10360 PIDs from Tracking System
This was done programmatically, comparing JSON output at
https://tracksys.lib.virginia.edu/admin/components/500020/tree.json
against the tree structure of this EAD XML document.

I intentionally skipped sections where the number of c0{n} nodes did
not match the number of objects in the relevant part of the JSON, as
there are a few locations where either the XML or the JSON has information
lacking in the other.  These should be examined by a knowledgeable party
and added manually.  Cursory analysis of the gaps indicates the scope
of the variations is a only about a dozen instances.