crude printouts of the tree structure of viuh00010.xml and the
JSON output of the Tracking System data for this same collection.
Use of 'diff' is a quick and easy way of spotting how these differ.
Numbers and spacing are an attempt to indicate where in a given branch
an object lies.  These can be used in XPath to find relevant XML.